merge main

+++ app/src/main/java/com/fpt/team5/golddigger/TransactionActivity2.java
@@ -101,12 +101,13 @@ public class TransactionActivity2 extends AppCompatActivity {
     }
 
     private void onBtnAddClick(View view) {
+
         if (edtAmount.getText().toString().equals("")) {
             Toast.makeText(this, "Please fill amount fields", Toast.LENGTH_SHORT).show();
         } else {
             String title = edtTitle.getText().toString();
             String description = edtDescription.getText().toString();
-            float amount = Float.parseFloat(edtAmount.getText().toString());
+            double amount = Double.parseDouble(edtAmount.getText().toString());
             String date = dateTimePickerCreate.getText().toString();
             int userId = pref.getInt("userId", 0);
             int categoryId = context.getCategoryByName(category);
